feat: Migración completa a Spring Boot 3.5.6
 MIGRACIÓN DAO  SPRING BOOT COMPLETADA
==================================================

 ARQUITECTURA MIGRADA:
 DAO manual  Spring Data JPA Repositories
 Gestión conexión manual  HikariCP automático
 Inicialización BD manual  DatosIniciales @ApplicationRunner
 Tests JUnit simples  @SpringBootTest con contexto Spring

 ESTRUCTURA REORGANIZADA:
 Paquetes: com.clinicaodontologica.{entity, repository, service, config}
• Entidades JPA: @Entity, @Table, @Id, @GeneratedValue
 Repositorios: JpaRepository<Entity, Integer> con métodos auto-generados
 Servicios: @Service con @Autowired para inyección de dependencias

 BASE DE DATOS:
 H2 en ~/clinicaFeliz (compatible con versión original)
 Hibernate DDL auto-generado equivalente a BD.crearTablas()
 Mismos datos de prueba: Homer/Marge + Dr. Juan/Dra. Maria

 TESTS MIGRADOS:
 @SpringBootTest + contexto completo Spring
 Lógica DADO-CUANDO-ENTONCES preservada
 application-test.properties para configuración tests

 FUNCIONALIDADES:
 Aplicación Spring Boot ejecutable con embedded Tomcat
 H2 Console disponible en /h2-console
 Logs SQL para debugging
 Validaciones de negocio mantenidas (email único, matrícula única)

 CAMBIOS PRINCIPALES:
+ Spring Boot 3.5.6 parent POM
+ Spring Data JPA repositories
+ Entidades JPA con anotaciones
+ ApplicationRunner para datos iniciales
+ Tests con contexto Spring
+ Configuración centralizada
- Clases DAO manuales eliminadas
- BD.java y gestión manual conexiones
- Tests sin contexto Spring

 PRÓXIMOS PASOS:
 [ ] REST Controllers (@RestController)
 [ ] Vistas web (Thymeleaf templates)

==================================================
Sistema completamente funcional y listo para uso

+++ src/main/java/com/clinicaodontologica/service/OdontologoService.java
@@ -46,7 +46,7 @@ public class OdontologoService {
         return odontologoRepository.save(odontologo);
     }
 
-    public Odontologo buscarOdontologoPorId(Long id) {
+    public Odontologo buscarOdontologoPorId(Integer id) {
         if (id == null || id <= 0) {
             System.out.println("Error: ID debe ser válido");
             return null;
@@ -75,7 +75,7 @@ public class OdontologoService {
         return odontologoRepository.save(odontologo);
     }
 
-    public void eliminarOdontologo(Long id) {
+    public void eliminarOdontologo(Integer id) {
         if (id == null || id <= 0) {
             System.out.println("Error: ID debe ser válido");
             return;
